Moved the init wrapper method and its helpers into a separate function, these were making the context file a little monolithic.
